bug resolving api feedback [finish#168899682]
