+++ server/routes/userRoutes.js
@@ -1,25 +1,20 @@
 import express from 'express';
-import Users5 from '../controllers/ViewSpecific';
-import Users3 from '../controllers/Comment';
-import Users from '../controllers/signup';
-import Users1 from '../controllers/signin';
-import Users2 from '../controllers/createArticle';
-import Users7 from '../controllers/edit';
-import Users4 from '../controllers/deleteArticle';
-import Users6 from '../controllers/viewAll';
+import {signup,signin} from '../controllers/user';
+import {editanarticle,createanarticle,deleteanarticle,viewspecificarticle,
+    viewallarticle, commentonarticle } from '../controllers/Articles';
 import validate from '../helpers/validations';
-import auth from'../helpers/middlewares';
+import auth from'../controllers/auth';
 const router = express.Router();
 
-router.post('/auth/signup', validate.signup, Users.signup);
-router.post('/auth/signin', validate.signin, Users1.signin);
-router.post('/articles',auth, validate.createanarticle, Users2.createanarticle);
-router.patch('/articles/:id',auth, Users7.editanarticle);
-router.delete('/articles/:id',auth, Users4.deleteanarticle);
-router.post('/articles/:id/comments', auth, validate.commentonanarticle,Users3.commentonanarticle );
+router.post('/auth/signup', validate.signup, signup);
+router.post('/auth/signin', validate.signin, signin);
+router.post('/articles',auth, validate.createanarticle,createanarticle);
+router.patch('/articles/:id',auth, editanarticle);
+router.delete('/articles/:id',auth,deleteanarticle);
+router.post('/articles/:id/comments', auth, validate.commentonanarticle, commentonarticle );
 
 
-router.get('/feeds',auth,Users6.viewallarticle);
-router.get('/articles/:id',auth,Users5.viewspecificarticle);
+router.get('/feeds',auth,viewallarticle);
+router.get('/articles/:id',auth,viewspecificarticle);
 
 export default router;
